64bit fixups
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/branches/UNICODE@4948 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/exehead/fileform.h
@@ -250,7 +250,7 @@ typedef struct
 
 // nsis blocks
 struct block_header {
-  int offset;
+  ptrdiff_t offset;
   int num;
 };
